Move formatExtendedField and formatIfListAndRemoveDuplicates functions into util.js and renderUtils.jsx accordingly
